Refactor exception handling for Atlassian integration in 3lo-token-broker.
- Split `GlobalExceptionHandler` into specialized handlers: `AtlassianExceptionHandler` and `GeneralExceptionHandler`.
- Introduce `ExceptionService` for reusable exception response logic.
- Migrate Atlassian-specific exceptions to a dedicated `exception.atlassian` package.
- Update related classes and services to reflect new structure.
