+++ src/main/java/com/solesonic/exception/atlassian/AtlassianTokenException.java
@@ -1,22 +1,24 @@
-package com.solesonic.exception;
+package com.solesonic.exception.atlassian;
+
+import org.springframework.http.HttpStatusCode;
 
 public class AtlassianTokenException extends RuntimeException {
-    private final String errorCode;
+    private final HttpStatusCode errorCode;
     private final boolean retriable;
 
-    public AtlassianTokenException(String message, String errorCode, boolean retriable) {
+    public AtlassianTokenException(String message, HttpStatusCode errorCode, boolean retriable) {
         super(message);
         this.errorCode = errorCode;
         this.retriable = retriable;
     }
 
-    public AtlassianTokenException(String message, String errorCode, boolean retriable, Throwable cause) {
+    public AtlassianTokenException(String message, HttpStatusCode errorCode, boolean retriable, Throwable cause) {
         super(message, cause);
         this.errorCode = errorCode;
         this.retriable = retriable;
     }
 
-    public String getErrorCode() {
+    public HttpStatusCode getErrorCode() {
         return errorCode;
     }
