+++ src/main/java/com/solesonic/exception/atlassian/JiraExceptionResponse.java
@@ -1,4 +1,4 @@
-package com.solesonic.exception;
+package com.solesonic.exception.atlassian;
 
 public record JiraExceptionResponse(String uri, String response) {
 }
