+++ src/main/java/com/solesonic/exception/atlassian/RefreshTokenConflictException.java
@@ -1,4 +1,4 @@
-package com.solesonic.exception;
+package com.solesonic.exception.atlassian;
 
 public class RefreshTokenConflictException extends RuntimeException {
     public RefreshTokenConflictException(String message) {
